Extract out texture animation export to its own module
